requirements.txt
python-dotenv
psycopg2-binary
google-auth 
google-auth-httplib2 
cloud-sql-python-connector
pg8000
pandas
numpy
xgboost
scikit-learn==1.5.2